+++ app/controllers/users_controller.rb
@@ -1,7 +1,7 @@
 class UsersController < ApplicationController
 
-  before_action :set_user, only: [:show, :update, :destroy]
-  skip_before_action :get_current_user, only: [:index, :show, :create]
+  before_action :set_user, only: [:show, :update, :destroy, :seller_score, :buyer_score]
+  skip_before_action :get_current_user, only: [:index, :show, :create, :seller_score, :buyer_score] 
 
   def index 
     render_ok User.all
@@ -27,11 +27,33 @@ class UsersController < ApplicationController
     end
   end
 
+  def seller_score
+    sold_products = @user.sold_products
+    if (score = sold_products.inject{ |sum, element| sum + element }.to_f / sold_products.size).nan?
+      render_ok 0
+    else 
+      render_ok score
+    end
+  end
+
+  def buyer_score
+    bought_products = @user.bought_products
+    if (score = bought_products.inject{ |sum, element| sum + element }.to_f / bought_products.size).nan?
+      render_ok 0
+    else 
+      render_ok score
+    end
+  end
+
   def destroy
-    if @user.id == @current_user.id
-      render_ok @current_user.destroy
-    elsif is_current_user_admin.nil?
-      render_ok @user.destroy
+    if @current_user.sold_products.empty? and @current_user.bought_products.empty? 
+      if @user.id == @current_user.id
+        render_ok @current_user.destroy 
+      elsif is_current_user_admin.nil?
+        render_ok @user.destroy 
+      end
+    else
+      render json: {authorization: 'we have to preserve the history of the web page'}, status: :unprocessable_entity
     end
   end
